backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-watson
drf-spectacular
celery
django-celery-results
redis
Pillow
faker